# backend-go/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ClaimItemListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.ClaimItem'
        type: array
      total:
        type: integer
    type: object
  dto.CreateClaimItemRequest:
    properties:
      cost:
        minimum: 0
        type: number
      faulty_part_id:
        type: string
      issue_description:
        maxLength: 1000
        minLength: 10
        type: string
      part_category_id:
        type: string
      replacement_part_id:
        type: string
      type:
        type: string
    required:
    - cost
    - faulty_part_id
    - issue_description
    - part_category_id
    - type
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.SuccessResponse:
    properties:
      data: {}
    type: object
...
  entity.ClaimItem:
    properties:
      claim_id:
        type: string
      cost:
        type: number
      created_at:
        type: string
      faulty_part_id:
        type: string
      id:
        type: string
      issue_description:
        type: string
      part_category_id:
        type: string
      replacement_part_id:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ev-warranty.com
    name: API Support
  description: API for EV Warranty Management System
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EV Warranty API
  version: "1.0"
paths:
...
  /claims/{id}/items:
    get:
      consumes:
      - application/json
      description: Retrieve all items for a specific claim
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Claim items retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ClaimItemListResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Get claim items by claim ID
      tags:
      - claim-items
    post:
      consumes:
      - application/json
      description: Add a new item to a claim (SC Staff only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Claim item creation data
        in: body
        name: createClaimItemRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CreateClaimItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Claim item created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.ClaimItem'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new claim item
      tags:
      - claim-items
  /claims/{id}/items/{itemID}:
    delete:
      consumes:
      - application/json
      description: Remove an item from a claim (SC Staff only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Claim Item ID
        in: path
        name: itemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Claim item deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim item not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete a claim item
      tags:
      - claim-items
    get:
      consumes:
      - application/json
      description: Retrieve a specific claim item by its ID
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Claim Item ID
        in: path
        name: itemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Claim item retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.ClaimItem'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim item not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Get claim item by ID
      tags:
      - claim-items
  /claims/{id}/items/{itemID}/approve:
    post:
      consumes:
      - application/json
      description: Approve a claim item for processing (EVM Staff only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Claim Item ID
        in: path
        name: itemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Claim item approved successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim item not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Approve a claim item
      tags:
      - claim-items
  /claims/{id}/items/{itemID}/reject:
    post:
      consumes:
      - application/json
      description: Reject a claim item (EVM Staff only)
      parameters:
      - description: Claim ID
        in: path
        name: id
        required: true
        type: string
      - description: Claim Item ID
        in: path
        name: itemID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Claim item rejected successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Claim item not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Reject a claim item
      tags:
      - claim-items
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"